[Fix] ChangeName
PJW 이름 붙은 클래스 이름 변경
ex) PJWTableManager -> TableManager
