--- Assets/BIK/02.Scripts/Table/TableBase.cs
@@ -1,8 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public abstract class TableBase
-{
-    public abstract void Load();
-}
