--- Assets/BIK/02.Scripts/Table/TableBase.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 43df966958ee8664ba90ddfdfc45a7f9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
