--- Assets/BIK/02.Scripts/Table/TableManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: cccee570a68306745a4ccbe055e8977c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
