--- Assets/PJW/Prefab/Items.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 33665cb505d5eef419887fd775784189
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
