+++ Assets/PJW/Script/Table/PJWTableBase.cs
@@ -2,6 +2,6 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public abstract class PJWTableBase
+public abstract class TableBase
 {    public abstract IEnumerator Load();
 }
