+++ Assets/PJW/Script/Table/PJWTableType.cs
@@ -2,7 +2,7 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public enum PJWTableType
+public enum TableType
 {
     Item
 }
